Update imports and naming for fotografie and thought sections
